Jindra/export multiple surfaces

+++ src/geometry/ColorManager.h
@@ -1,9 +1,10 @@
 #pragma once
 
-#include <random>
-#include <vector>
 #include "cinder/Color.h"
 #include "glm/glm.hpp"
+
+#include <random>
+#include <vector>
 
 #define PEPR3D_MAX_PALETTE_COLORS 8
 
@@ -38,7 +39,7 @@ class ColorManager {
         std::mt19937 gen(rd());  // Standard mersenne_twister_engine seeded with rd()
         std::uniform_real_distribution<> randomGen(0.0, 1.0);
 
-        for(int i = 0; i < number; ++i) {
+        for(size_t i = 0; i < number; ++i) {
             mColorMap.emplace_back(randomGen(gen), randomGen(gen), randomGen(gen), 1);
         }
     }
